feat: create @jahia/vite-plugin
