+++ vite-plugin/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@jahia/vite-plugin",
+  "version": "0.5.0-SNAPSHOT",
+  "repository": "git@github.com:Jahia/javascript-modules.git",
+  "license": "MIT",
+  "type": "module",
+  "exports": {
+    "types": "./dist/index.d.ts",
+    "import": "./dist/index.js"
+  },
+  "scripts": {
+    "build": "pkgroll"
+  },
+  "devDependencies": {
+    "@types/node": "^22.13.5",
+    "pkgroll": "^2.11.0",
+    "vite": "^6.1.1"
+  },
+  "peerDependencies": {
+    "vite": ">=6.0.0"
+  }
+}
